Second Commit: Introduction of base for Collision engine.
The Beginnings of a Collision engine was added, various changes to
existing classes, namely the GameObject interface. The addition of
a World interface and an AbstractWorld class in which GameObjects
will exist during program runtime. Some JavaDoc was added to
AbstractWorld and the GameObject and Solid interfaces have been
Documented completely. Height and Width fields have been added to
AbstractGameObject and their respective accessor and mutator methods
have been added to GameObject and Overriden in AbstractGameObject.

Signed-off-by: Mihail Kurbako <Mihail@.(none)>