Physics engine commit: Gravity/Magnetism implemented, Slick2D.
This commit brings changes to the Movable interface, allowing for both
speed (by adding mutators/accessors for both x and y speed) and for magnetic
and gravitational forces. The AbstractMovable object was added as well, to
implement the new methods, as well as to reduce redundant coding in the sub
classes. The redundant code was removed from said classes. The GameObject
interface also received an update in the form of a new accessor for magnetic
strength, and a corresponding variable in AbstractGameObject. In other news
the Eclipse Project enclosed contains functional Slick2D, set up to run on
any system, excluding Solaris and Solaris-based Operating Systems.

Signed-off-by: Dan Zapronikov <danzap95@gmail.com>