Minor Rework
Not truely a Commit. Just an addition for the basic framework
for the soon to be added LimitedWorld class, functionality
for recalculateMovement and handleMovement methods. The addition
of a bit of JavaDoc, and most importantly, the Movable interface
is now extends the GameObject interface.
TODO: Dan: Implement methods to calculate displacement of an Object
based on its velocity.

Signed-off-by: deltamanx <bobvonjoe@gmail.com>