Threading and Documentation
The title summarizes what I have done. As a note, I have removed the move
method as movement is handled by the Engine as it needs to check for
collsions. I added a MainProgram for testing purposes and have refined the
Threads so that they function properly. Some slight modifcations here and
there. Created seperate methods to countXMagnetism and countYMagnetism,
which are invoke within countMagnetism [for purposes of code readability].

Signed-off-by: deltamanx <bobvonjoe@gmail.com>