Phys engine update
This commit changes the speed system, making the speed pixels/second, and
changes how the game updtes locations to reflect that change. Slight
adjustments have been made to the executable class for testing, and some
other minor changes have been made, such as adding a dat folder for storing
non-code files, as well as a tweak to the magnetic constants.

Signed-off-by: DanSze