States and Menus.
Added states support. This will be the driving force behind the gameplay.
Each screen is a state that exists seperately form the other states. The
main game can focus on any of the states, and switch between them freely.
States remain unchanged between switches, so they have to be manualy reset
when one wishes to start from a fresh screen. See the slick wiki for more
information. In other news, it seems that bouncing is still not working as
intended, sometimes switching BOTH directions.

Also, a reminder to start drawing screen backgrounds, propper images, and the like.

Signed-off-by: DanSze <danzap95@gmail.com>