Large update. Changed several things to do with graphics, such as target. Added arrow that points and scales with your launch vector. Added a level selection screen (more on that in a bit). Added help screen.
The selection screen will automatically update with every level put into
the level data folder, and currently lists the difficulty number and the
stage number. Later, when we have several levels, it will show a screenshot
of the screen, the name of the level, and navigating instructions. This will
most likely be in the for of a single image, except the instructions.

Also, reminder to start working on highscores, since I have no idea how you
will handle the local saving of highscores and the online highscores (seeing
as Ms. Dyke asked for the former and you wanted the latter).

I'll get on making the levels.

Signed-off-by: DanSze <danzap95@gmail.com>