Stage Select Complete
Stage selection now shows a preview of the stage, as well as a stae name.

When you add a new stage, remember to add a name to the name array in the
StageSelectState class. The 2d array is 3x10, with the first being difficulty
(0-2) and the second being the stage number (0-9).
Remember that the name difficulty number is staggered by one.

Also of note is the three new stages added.

Signed-off-by: DanSze <danzap95@gmail.com>